fix(mobile): repair l10n config, missing settings stub, deprecated APIs
- l10n.yaml: remove deprecated synthetic-package, add output-dir
- Add features/profile/settings_screen.dart stub (router import)
- app_theme.dart: CardTheme -> CardThemeData
- auth_interceptor.dart: QueuedInterceptors -> QueuedInterceptorsWrapper
- splash_screen.dart: cast (double - int) to int for Duration

+++ mobile/lib/features/auth/presentation/splash_screen.dart
@@ -30,7 +30,7 @@ class _SplashScreenState extends ConsumerState<SplashScreen> {
     final elapsed = DateTime.now().difference(start).inMilliseconds;
     if (elapsed < AppConstants.splashDelayMs) {
       await Future<void>.delayed(
-        Duration(milliseconds: AppConstants.splashDelayMs - elapsed),
+        Duration(milliseconds: (AppConstants.splashDelayMs - elapsed).toInt()),
       );
     }
